Protect fragile parbox in demothesis author field
On some older versions of LaTeX, an error occurs when trying to use a
"\parbox" within the author field of the demothesis:
  ! Argument of \reserved@a has an extra }.

For example, the error occurs on Ubuntu 14.04 LTS when using the LaTeX
packages available in default repositories (i.e. TeX Live 2013/Debian).
It doesn't seem to occur on Ubuntu 16.04 LTS (TeX Live 2015/Debian).

For more information about this kind of error, do a web search for:
  Use of fragile command in a moving argument

Signed-off-by: Martin Erola <marer248@student.liu.se>